completed migration of layesr into s3 and refactored file structure and import statement to reflect aws lambda
