+++ src/extract.py
@@ -2,14 +2,14 @@ import logging
 import boto3
 from botocore.exceptions import ClientError
 from io import BytesIO
-from utils.db_utils import connect_to_db, close_db_connection
-from utils.utils import upload_file
+from db_utils import connect_to_db, close_db_connection
+from utils import upload_file
 import polars as pl
 
 from datetime import datetime
 
 
-def lamda_handler(event, context):
+def lambda_handler(event, context):
     s3_client = boto3.client("s3")
     response = extract_data(s3_client)
     return response
@@ -67,5 +67,3 @@ def extract_data(s3_client=None, bucket="ainsdale-ingestion-bucket"):
         close_db_connection(conn)
 
 # remove this going forward
-if __name__ == '__main__':
-    lamda_handler(None, None)
